rib_management app: poll agent regularly, no RRC measurements app (rrc_trigger)
the RibManagement app: polls the agent regularly. Its first task is to give the
user some feedback if an agent cannot be reached some time and second to cause
a disconnect message if the connection to the agent is broken by periodically
polling the agent (this occurs if no other messages are send to the agent).

This app is not necessary for the disconnect message functionality itself.

The RRC measurements app is disabled because the threading model creates a new
thread for every app call, incurring high load. Therefore, to keep the old
load, this app is replaced by RibManagement.