Workaround: use mutexes in ue_mac_rib_info and enb_rib_info
use mutexes in enb_rib_info for {eNB,ue,lc}_config_ and in ue_mac_rib_info for
mac_stats_report_mutex_

the REST north bound runs concurrently with the RIB (RIB writes, REST
northbound reads). For a lot of reads, this leads to races, so mutexes are used
to mutually exclude both accesses. Initially, this has not been foreseen in the
thread model, running first the RIB and then the Apps.

A better way would probably to block all reads in the RIB while it is writing.
This could be implemented with a mutex, a condition_variable and a bool or by
blocking the stats_call app while the RIB is active.