add helper function for guessing firstRb
the agent checks that UL slices do not overlap. In order for maintaining the
short version of the slice configuration end point, the first_Rb has to be
added. This commits adds a helper function that adds a first_Rb. If this call
does not work, a first_rb value has to be specified manually