Stats_manager: save readable per-BS stats requests
This updates the stats_manager app to save per-BS statistics requests in
a human-readable format. The protobuf definition is changed to contain a
list of cell/UE-reports (instead of only bit-wise flags, to be used in
a JSON) and creates a new message, representing a list of reports. Then,
a new method for a default report is added, and when a new BS connects,
the currently active statistics requests is saved.